chore(backend): dead code, import cycle, structured logging, unused dep
Tier 4 architecture review (PR1) — backend hygiene.

- Delete the dead WikiAdapter.fetchPageContentWithSlots (no production
  callers, and it violated the error contract by returning "" instead of
  throwing on failure) plus its test and the server.test mock.
- Break the vintedSession <-> browserPrime import cycle: parseSetCookie
  moves to a new services/cookies.ts (re-exported from vintedSession for
  compatibility; browserPrime imports from cookies).
- Replace console.* with the structured createLogger in
  duplicateSyncService (drop 2 debug lines, 1 → log.info) and
  lpSyncService (log.error).
- Drop the unused WikiAdapter dependency from DuplicateSyncService
  (constructor + syncManager wiring + test).

Suite 473 green (-1 = removed the dead-method test), lint clean, build OK.
