+++ services/vintedSession.ts
@@ -3,29 +3,13 @@ import https from "https";
 import { getRandomUserAgent } from "../scrapingClient";
 import { vintedRequestHeaders, VintedSession } from "./vintedHttp";
 import { primeWithBrowser } from "./browserPrime";
+import { parseSetCookie } from "./cookies";
+
+// Re-exported for existing importers/tests; the implementation lives in `cookies.ts`
+// so `browserPrime` can use it without a `vintedSession ↔ browserPrime` cycle.
+export { parseSetCookie };
 
 const VINTED_HOME = "https://www.vinted.pl/";
-
-/**
- * Assembles the `Cookie` header from a response's `Set-Cookie` array (axios `res.headers["set-cookie"]`).
- * Takes only the `name=value` pair (the part before the first `;`), skips attributes (Path/Expires/…)
- * and empty/malformed entries. Pure function — easy to test.
- */
-export function parseSetCookie(setCookie: string[] | undefined | null): string {
-  if (!Array.isArray(setCookie)) return "";
-  const pairs: string[] = [];
-  const seen = new Set<string>();
-  for (const raw of setCookie) {
-    if (typeof raw !== "string") continue;
-    const pair = raw.split(";")[0]?.trim();
-    if (!pair || !pair.includes("=")) continue;
-    const name = pair.slice(0, pair.indexOf("=")).trim();
-    if (!name || seen.has(name)) continue;
-    seen.add(name);
-    pairs.push(pair);
-  }
-  return pairs.join("; ");
-}
 
 /**
  * „Warming up" a Vinted session: one GET of the home page with a real browser signature,
